backend/requirements.txt
Flask
Werkzeug
torch
torchvision
transformers
Pillow
tensorflow
pandas
numpy